pyproject.toml
[project]
name = "ome-hackathon-season-14"
version = "0.1.0"
description = "OME Hackathon Season 14 - Text analysis microservices for media ecology"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core API dependencies
    "fastapi[standard]>=0.128.0",
    "httpx>=0.27.0",
    "uvicorn>=0.40.0",
    # Database
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.46",
    # Data processing
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    # NLP & ML microservices
    "rake-nltk>=1.0.6",
    "spacy>=3.7.0",
    "torch>=2.10.0",
    "transformers>=5.0.0rc3",
    # Analysis notebooks (optional)
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "wordcloud>=1.9.6",
]
